ResultPage displays entries depending on isDetailed key of item.json
